microservice-python/requirements.txt
pdfplumber
pytesseract
Pillow
requests
pdf2image
openpyxl